[devbox] Add anonymized telemetry (#40)
## Summary
Add anonymized telemetry to understand usage of devbox and further improve it. We allow users to opt-out via an environment variable.

To add telemetry I implemented a small "midcobra" framework, that make it possible to add "middlware" to cobra CLIs. The plan is to re-use the midcobra functionality in other binaries like `envsec`.

## How was it tested?
Built locally, ran, and checked telemetry logs.