[devbox] Allow for custom install/build/start commands (#43)
## Summary
Allow for custom install/build/start commands by customizing them in `devbox.json`. This makes it possible for users to build docker images in the cases when they need to do some customization, or when we have not yet written a language planner for that language.

## How was it tested?
Built, and ran `devbox plan` and `devbox generate` against a few examples.