[RFC] move gitignore file to .devbox directory and ignore gen/profile/shell_history (#215)
## Summary

With the recent use of nix-profile, the `profile` symlinks are showing
up in git history.
This is regrettable.

This PR adds a `.gitignore` file to `.devbox` directory and ignoring the
`gen/`, `profile*`
, `nix-profile` and `shell-history` symlinks.

Going forward, we should avoid editing this file. Each edit will need to
be checked
in by our users. So, this PR is a breaking change, in a sense.

New features we add that generate files in `.devbox` should be added
inside `gen/`
if they want to be gitignored.

New features that we do want users to check in (like a lock file) can be
placed
in `.devbox/` (and not `.devbox/gen`).

## How was it tested?

in a new devbox project, with `git init` and initial files already
committed:
- did `devbox shell` to generate the gitignore and other files 
- did `git status` to inspect changes and found no new file changes were
reported.
- did `ls -al .devbox` and saw the other files (like gen/ and profile*
were generated) as well as `.devbox/.gitignore`.