nix: fix removing user Nix profiles from PATH (#220)
When launching a devbox shell, we filter out any Nix user profiles from
the PATH (i.e., ~/.nix-profile). This fixes a couple of bugs with how we
determine if a PATH directory matches a Nix profile:

1. The `strings.HasPrefix("/Users", profileDir)` call is backwards.
2. Checking for `/Users` is macOS-specific and won't work on Linux.
Instead, check if the directory starts with `/nix` to determine if it's
the default system profile.