Point the cli to the production gateway (#327)
## Summary
Point the CLI to the production gateway (`gateway.devbox.sh`) by
default.

Note that this means that if you're working on the gateway, you'll need
to re-compile the devbox binary to test against your development
version. Let me know if you prefer I add an (internal) flag to select
which gateway to talk to.

## How was it tested?
Didn't.