cloud: use unique SSH keys per VM (#340)
When the gateway creates a new VM for a user, it also generates an
ed25519 key pair that the client can use to login. This change updates
the client to save the key pair in `~/.config/devbox/ssh/keys` and
configures SSH to use it when connecting to the VM.

- Consolidate `cloud/sshclient` and `cloud/openssh` into `cloud/openssh`
to make it clearer that they manipulate the OpenSSH command.
- Remove `openssh.Client.Port` and instead parse the port from the
hostname. This allows the port to be overridden via `DEVBOX_GATEWAY`.
- Add `openssh.AddVMKey(host, key)` that adds a new key to
`~/.config/devbox/ssh/keys`.
- Atomically edit `~/.ssh/config` and `~/.config/devbox/ssh/*` files so
that partial edits don't break the user's SSH.
- Add a bunch of tests that check that the client correctly modifies the
user's `~/.ssh/config` and `~/.config/devbox` files.
- Misc. bug fixes related to SSH configs:
	- Quote paths in ~/.ssh/config to handle paths with spaces.
- Make the devbox include regex case-insensitive and detect when the
include is commented out.
	- Set recommended permissions on the various SSH directories and files.