[nix-installer] auto-install nix if we detect it's missing (#343)
## Summary

This change makes devbox attempt to install nix if it is missing and the command the user ran requires it. It shows yellow text alerting the user before it starts (otherwise it is really jarring). One improvement could be to combine the sudo warning into the yellow text.

If the installation fails we show the error that caused it. If the installation succeeds we show a user warning which is yellow text. This warning is an error which causes the CLI execution to end. We need this because nix requires restarting the terminal. This is another area where having a pkg installer may be an improvement. 

cc: @Lagoja 
## How was it tested?

Ran `devbox add mariadb` without having nix installed. Got:

<img width="471" alt="image" src="https://user-images.githubusercontent.com/544948/205784786-c76d90f2-9846-418e-93db-d2605eddfc32.png">

Pressing enter started the installation.

After installation we show:

<img width="741" alt="image" src="https://user-images.githubusercontent.com/544948/205785207-da291c1b-2c5f-4004-9bb5-a3a3e053e94c.png">

Co-authored-by: Greg Curtis <greg.curtis@jetpack.io>