[cloud] ssh shim part 4: detect stopped vm, and terminate mutagen sessions (#352)
## Summary

This PR adds logic for the `sshshim` to check for a live VM. If the VM
is not reachable,
then it should terminate the sync session. We label each sync with the
machineID
so that cloud-shells to the same vm get the same label for their mutagen
sync.

To test for liveness, I currently execute a `echo "alive"` command in
the VM. There
are likely better ways to do this. Suggestions welcome!

NOTE:
For now, I pass-thru the ssh/scp command if the arguments indicate that
the server
is not a devbox-vm. In the next PR, I'll see if we can have a dedicated
mutagen daemon
via setting MUTAGEN_DATA_DIRECTORY, in which case we can more strongly
impose
that the ssh/scp command can be executed only with a liveVM. UPDATE:
this works!

NOTE:
I also discovered that I need to manually empty-out the MUTAGEN_PROMPTER
env-var.
This is done, because mutagen otherwise will expect a certain
(interactive?) prompt
in the arguments to the mutagen CLI. Mutagen does set this env-var on
all ssh/scp executions, but I'm a bit unclear on why this is hooked up.

## How was it tested?

0. ensure no active mutagen daemon (next PR should take away this step)

1. manually start a vm (command omitted due to closed source, but happy
to share)
2. `DEVBOX_FEATURE_SSH_SHIM=1 DEVBOX_VM=username@vmaddress devbox cloud
shell`
3. `cd ~/.config/devbox/ssh/shims` and `tail -f logs.txt` to monitor the
shim.
4. exit the cloud shell 
5. `fly m stop <id>` to stop the vm
6. `touch foo.txt` in the devbox project folder to trigger a sync action
7. wait for the `logs.txt` to show the `mutagen sync terminate` command.
8. `~/.cache/mutagen/bin/mutagen sync list` to confirm the session is
terminated.