[nix] nix install without using daemon flag DEV-1266 (#368)
## Summary

Previously we were using `--daemon` to force multi-user on all platforms. This doesn't work on WSL, so omit the flag. This means on macOS it will use multi-user and on linux it will install as single user.

## How was it tested?

Untested, but this is what it used to be and was tested previously.