cloud: change ControlMaster to auto and add sockets dir (#427)
## Summary

Setting ControlMaster to auto prevents a warning from being shown when
there's already a ControlMaster process running. Put the control sockets
in a separate directory in case we later support multiple VMs per user.

Also move the socket-specific code to the openssh package so that
hostname parsing lives with sshconfig.tmpl.

## How was it tested?

Unit tests and connecting to a VM multiple times.