[nix] attempt to source nix startup files if not in path (#426)
## Summary

This change allows devbox to work even if nix is not in path. This is useful for:

* Right after installation we no longer require restarting terminal
* After macOS updates (which clear `/etc` rc files) devbox should continue to work. 

I tried a bunch of other approaches, but this one seems like most robust and least prone to errors because it does't actually change the commands. If we find a `/nix` dir but `nix-shell` is not in path, we attempt to source the nix startup files (multi and single user). If we still fail to find the binaries we show an error.

Edit:

Additional changes:

* Added `-y` flag to avoid prompt
* Moved some stuff out of nix package so that it doesn't depend on cobra commands/flags

cc: @LucilleH 

## How was it tested?

- [x] exported a PATH that was missing nix paths. Confirmed that `nix version` did not work. Devbox continued to work.
- [x] Fresh instal on ubuntu does not require restarting terminal