Extract shell-specific init file injection into own func (#424)
## Summary

While this seems like a no-op, please note that I removed the quotes
around the formatted string in each of the cases. I was finding that the
rc files weren't being executed when using the quotes (and while using
nixless shell).

## How was it tested?
Add a var into `shellrc.tmpl`, build devbox, run `./devbox shell`, and
check if the var is defined. Do this for bash, zsh, ksh.