Adds code to call nix print-dev-env (#456)
## Summary
This simply adds the ability to call `nix print-dev-env -f <file>`, parses the result and returns it. The idea is to then use the output of this to implement a better (read faster and more isolated) version of `devbox run` and eventually `devbox shell`. The output can also be used for a more correct implementation of `devbox shell --print-env` (albeit slower).

## How was it tested?
Used it for experimental implementation of `devbox shell` that relies on it. Will send that on a separate PR.