[polish] Minor polish on help text (#475)
## Summary
To follow convention of (a) help for commands is capitalized, (b) help
for flags isn't.

I also removed "quiet mode" and "verbose", because when you print the
help, "quiet" and "verbose" is printed directly next to it, so it's kind
of redundant.

## How was it tested?
Ran `./devbox` and `./devbox help <cmd>` for every command