[cloud] Improve cloud shell error messages (#548)
## Summary

This shows better error messages when cloud shell fails. A possibly
common case is when a nix package doesn't exist or doesn't build
correctly. Ideally we don't end the shell and let the user fix it, but
in the meantime this shows a more relevant error message.

To implement this efficiently and extensibly, this uses an io.Writer to
collect pre-defined ssh session errors without explicitly storing the
output.

## How was it tested?

* Added a bad package to my devbox.json and did `devbox cloud shell`
* To test generic failures did `devbox cloud shell`, turned off my wifi
and waited for a timeout

<img width="1200" alt="image"
src="https://user-images.githubusercontent.com/544948/215221477-63380ba9-8fd1-440b-9056-32e585d2131d.png">