[flake] php extensions example project (#631)
## Summary

Added an example devbox-project that has a php extension package. This
has been useful for doing adhoc testing when developing the feature of
"flakes with php extensions". In a future PR, I'll add a testscript (and
possibly remove this one).

## How was it tested?

```
> devbox shell

# with package `php81Extensions.ds`
> php index.php
Original vector elements
idx: 0 and elem: hello
idx: 1 and elem: world
done

# without package `php81Extensions.ds` 
> php index.php
ds extension is not enabled
```