[nix] Simplify profile priority, make it match flake file (#770)
## Summary

I was implementing a progress indicator for global install, and the profile install retry logic was making it a bit complicated. I then realized that the order we were assigning to priorities (latter package takes precedence) is actually inverse to what flake `buildInputs` does (it uses earlier precedence). In practice, this means the profile generated by devbox and the flake generated by devbox would have different behaviors. This is not terrible but also not ideal. So I removed the sorting and added some documentation to ensure we always keep the same (local, global) order.

Finally, I tried out a slightly different approach which doesn't require a profile install with retries. By reading the profile manifest file we're able to see what the highest number existing priority (lowest priority)  is and just increment it to achieve the next lowest priority. I tried this up to numbers in the thousands there doesn't seem to be any limits. 

Some benefits of this approach:

* It's easier to implement progress indicators that read stdout/err in a goroutine. 
* Better performance (no retries needed)
* more deterministic
* Code is a bit simpler

Drawbacks:
* Depends on the manifest
* Needlessly assigns more priorities than are needed

@savil curious your thoughts on this?

## How was it tested?

```bash
devbox global add php82
devbox add php
devbox run "php -v"
```