[Lagobot 2.0] part 1: generate testscript for a single devbox example project (#799)
## Summary

**Why**
Our goal is to make each Devbox release higher-quality by ensuring that
all the devbox-projects in `examples/` are able to run.

**What**
We aim to do so by adding a `run_test` devbox script to each project in
`examples/`, and ensuring that it can run. This has the additional
benefit of making the `devbox.json` self-documenting for any user who is
reading the code in the `examples/` directory: they can refer to the
`run_test` to understand how to execute that example project.

**How**
To do so in our automated tests:
1. We will generate a simple testscript for each of these
devbox-projects that calls
`devbox run run_test`.
2. We copy the devbox project's files and directories into the
testscript's workdir.

As a proof of concept, this PR runs a testscript for the
`examples/development/go/hello-world` project. A future PR will
generalize this to all of the devbox projects in `examples/`

## How was it tested?

Ran with this command:
```
DEVBOX_DEBUG=0 go test -run TestExamples/development/go/hello-world -count 1 ./testscripts
```