[nix] Re-add profile to path (#882)
## Summary

Re-add profile to path to ensure we expose binaries exposed by `propagated-build-inputs`

Some packages like `curl` expose binaries using `propagated-build-inputs`. We need to make sure these are in path.

A better solution is to just revert to using profile to create all the wrappers. This requires some work to avoid breaking php and Haskell because they do some pre-processing in the top level flake that would not be in the files installed in the profile. 

## How was it tested?

```
devbox shell
which curl # shows system curl
devbox add curl
hash -r
which curl # shows profile curl
```