[bin-wrappers] Use shellenv command in binwrappers (#876)
## Summary

This changes bin wrappers to use `eval $(devbox shellenv
--use-cached-print-dev-env)`

### Motivation:
For performance reasons we were printing out the shellenv results in the
bin wrappers. The problem with this approach is that the env vars in the
wrappers overwrite whatever the user sets. For example:

`USER=foo node -e "console.log(process.env.USER);"`

will print `mike` because `export USER="mike"` is in the wrapper.

### Solution:

* Revert to using `eval $(devbox shellenv)` in the wrapper.
* For performance, add hash to environment and skip evaluation if the
hash matches (this means shellenv has already been eval-ed).
* For performance add new `--use-cached-print-dev-env` flag to shellenv
that uses a cached copy of `nix print-dev-env` if it exists. The cache
is regenerated every time `computeNixEnv` is called so it happens on
shell, run, adding/removing package, when wrappers are created, and also
when shellenv is called without the flag.
* For performance added DO_NOT_TRACK to shellenv call in wrapper. This
can be removed once segment moves to queuing instead of blocking.

Latency impact when hash matches: None, or maybe a bit < 1ms faster.
Latency impact when hash doesn't match: ~20ms slower

Potential additional optimization 1: skip launcher when calling `devbox
shellenv`. This should reduce about ~10ms of latency.
Potential additional optimization 2: when hash matches, instead of
saving a wrapper we could save a symlink. This would eliminate extra
latency of wrapper.

## How was it tested?

* Tested by adding and removing packages while in shell.
* Timed calls to `hello` 
* Existing unit tests that test shellenv, bin wrappers, etc