[example tests] generate "set -e" to exit scripts upon failure (#869)
## Summary

RFC: we should have failures in the devbox-scripts reflect in a non-zero
exit status
from `devbox run <script>`.

There is some risk that customers whose code was previously running will
now fail. However, their scripts were indeed failing and its just that
this failure will now be exposed.

Update: Based on John's feedback in a comment below I am adding a
featuregate that is activated in examples tests. We'll make a plan for
rolling this out to users, in a way that is not a big surprise for them.

TODO:
- [x] fix examples which have failures that were previously masked. Done
in previous PRs in the stack.

## How was it tested?

Removed `curl` from `lepp-stack` packages, and ran:
```
DEVBOX_EXAMPLE_TESTS=1 DEVBOX_DEBUG=0 go test -count 1 -v -run TestExamples/stacks_lepp ./testscripts/...
```
This failed, demonstrating that the `set -e` was working as intended.

I then added the `curl` back and the above command passed.