[plugin] Add `php.ini` in the PHP plugin (#1039)
## Summary

Adds an editable `php.ini` to the PHP plugin. This is useful when
developers need to configure their PHP instance, and by creating it in
`devbox.d` we let developers customize it for each of their PHP
projects. This addresses #1035 and the issue reported by @jrseliga

## How was it tested?

* On a clean PHP project, run `devbox add php@latest` and verify that
the php.ini file is created in devbox.d
* Also check the updated example in this PR