[global] Add update/services/run commands to global (#1046)
## Summary

This adds:

```bash
devbox global update
devbox global services [subcommand]
devbox global run
```

commands to global. 

I fixed a few issues:

* Some commands were missing `--config` flag
* We were using `IsDevboxShellEnabled` instead of `IsEnvEnabled`. The
first (obsolete?) command only looks if we're in a devbox shell while
the second returns true if we are in an environment created for the
specific config.

I found bug that I didn't fix in this PR which is that if a package with
plugins/services is installed the environment must be refreshed. If done
with while `IsEnvEnabled` is false this is not an issue. This can be
fixed by sourcing the environment prior to running the service with
special care not to override user set variables.

cc: @Lagoja 

## How was it tested?

```bash
devbox global add apache
eval "$(devbox global shellenv)" # due to env bug
devbox global services up
```