[shellenv] filter out buildInput paths from PATH (#1091)
## Summary

**problem**
Doing `devbox rm <package>` within a devbox shell will usually still
have the package's binary accessible in PATH. This is confusing and
unexpected.

**cause**

In shellenv's `PATH`, we now have 3 sources of paths:
1. `.devbox/virtenv/.wrappers/bin`
2. `.devbox/nix/profile/default/bin`
3. `/nix/store/<hash>-<package>-<version>/bin`

When we do `devbox rm <package>`, the first two (wrappers and profile
bins) are removed, but the `/nix/store` remains in PATH.

The user would have to do `eval $(devbox shellenv)` to remove it. But
that is not a good UX.

**background context**

Heres’ why we need/don’t need each one:
1. bin-wrapper dir: is our path for binaries. Binaries here are wrapped
to ensure the correct environment is sourced. “Support files” are also
sym linked here to ensure that binaries that require relative
dependencies work as well (e.g. mariadb)
2. profile bin path: This is a fallback hack for certain edge cases
(e.g. curl). If a package uses propagateBuildInputs then we may not
correctly create a bin wrapper for it. Currently we use $buildInputs to
determine binaries that need to be wrapped, but this does not contain
propagated inputs, only explicit inputs.
3. nix store paths: these are included by nix from `nix print-dev-env`.
We've never removed them, but need to ensure nothing breaks.

Thanks to @mikeland86 for explaining this. I've copy-pasted this from a
chat conversation.


**this PR**
This PR removes from PATH the subset of `/nix/store` paths that belong
to the buildInputs, during `shellenv`.


Fixes #1060 

TODO:
- [x] do more extensive testing

## How was it tested?

```
> devbox add stress
> devbox shell
(devbox)> which stress
# printed wrappers/bin path
(devbox)> devbox rm stress
(devbox)> which stress
# empty
# BEFORE: this would print the /nix/store path
```

- [x] ensure testscripts pass.