[CLI] Changed default nix bash to bashInteractive (#1223)
## Summary
In `shellPath()` function where we decide which type of shell to use
when dropping user in devbox shell, if we can't figure out the shell, we
drop the user in the default `bash` that nix installs in store. But this
bash doesn't handle arrow keys and keyboard combinations very well, so
it creates a frustrating experience for users (including pure shell
users). `bashInteractive` package has all these capabilities and makes
the experience better for a 'default' shell.

So this PR changes this:
- `nix eval <bash>` to get store path and use to create default shell
to this:
- `nix eval <bashInteractive>` to get store path
- nix build <generated store path>
- use store path to create default shell

Addresses #1193

## How was it tested?
- compile
- ./devbox shell --pure
OR 
- unset $SHELL
- ./devbox shell