[plugins] Remove match and packages from schema (#1296)
## Summary

After chatting with @Lagoja we decided to remove `packages` and `match`
from plugin schema before we release custom plugins. The reason being:

* `packages` is poorly defined and not a great experience. It was added
so we could build plugins that install custom flakes (for php, haskell,
etc) and replace planners. It was renamed to `__packages` and will not
be externally documented. In very near future we plan on adding
`suggested_packages` (final name TBD) which informs the user of any
packages needed for a plugin and allows the user to install if needed.
* `match` is designed to automatically add built-in plugins if certain
packages are installed. This functionality doesn't make sense for custom
plugins. This field was moved into a hardcoded map. Another benefit of
removing this from schema is that we can put the plugins into the plugin
github repo.

cc: @Lagoja 

## How was it tested?

* unit tests 
*  manual smoke testing (installed php)